+++ src/main/java/roomescape/infrastructure/config/ClockConfig.java
@@ -1,4 +1,4 @@
-package roomescape.domain;
+package roomescape.infrastructure.config;
 
 import java.time.Clock;
 import org.springframework.context.annotation.Bean;
